Fix compiler warnings.
A missing return statement caused strange behaviour of a test case.
This and several similar cases have finally been found by gcc, and are
corrected, now. Other warnings include comparisons of signed and unsigned
integers, unused variables and wrong parenthesis.

+++ src/kernel/Execute.h
@@ -48,7 +48,7 @@ void MutResetKeys();
 void GlobalReset();
 
 void AddKey(mutabor_box_type * box, int taste, size_t id, size_t channel, void * userdata);
-void DeleteKey(mutabor_box_type * box, int taste, size_t id, size_t channel);
+void DeleteKey(mutabor_box_type * box, int taste, size_t id, int channel);
 void MidiAnalysis(mutabor_box_type * box, BYTE midiByte);
 
 void pascal _export KeyboardAnalyse(mutabor_box_type * box, int taste, char isLogic);
