+++ src/kernel/routing/gmn/DevGIS.h
@@ -164,7 +164,7 @@ namespace mutabor {
 			{};
 */
 
-		virtual int  GetChannel(int taste, int channel, int id)
+		virtual int  GetChannel(int taste, int channel, size_t id)
 			{
 				return -1;
 			}
